+++ packages/zowe-explorer-api/src/profiles/ProfilesCache.ts
@@ -165,6 +165,9 @@ export class ProfilesCache {
         let mProfileInfo: zowe.imperative.ProfileInfo;
         try {
             mProfileInfo = await this.getProfileInfo();
+            if (!mProfileInfo.usingTeamConfig) {
+                return;
+            }
             const allTypes = this.getAllProfileTypes(apiRegister.registeredApiTypes());
             allTypes.push("base");
             for (const type of allTypes) {